refactor: -reduce cognitive complexity - extract find-plugins in its own module
